+++ spring-data-elasticsearch-rest-starter/src/test/java/com/github/ydespreaux/spring/data/autoconfigure/elasticsearch/ITSuiteTest.java
@@ -37,6 +37,6 @@ import org.junit.runners.Suite;
 public class ITSuiteTest {
 
     @ClassRule
-    public static final ElasticsearchContainer elasticContainer = new ElasticsearchContainer("6.4.2");
+    public static final ElasticsearchContainer elasticContainer = new ElasticsearchContainer("6.4.3");
 
 }