+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 
 ## [1.1.0-SNAPSHOT]
 ### Added
-- Dependency elasticsearch version from 6.4.2 to 6.4.3 (o)
+- Dependency spring-boot version from 2.1.0 to 2.1.2
+- Dependency elasticsearch version from 6.4.2 to 6.4.3
 - Reactive repository (o)
 
 ### Changed